Incubator.Dialog - header refactor
