--- generatedTypes/src/incubator/Dialog/DialogDivider.d.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import { DialogHeaderProps } from './types';
-declare const _default: React.MemoExoticComponent<(props: Pick<DialogHeaderProps, "showDivider">) => JSX.Element | null>;
-export default _default;
