--- generatedTypes/src/incubator/Dialog/DialogKnob.d.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import { DialogHeaderProps } from './types';
-declare const _default: React.MemoExoticComponent<(props: Pick<DialogHeaderProps, "showKnob">) => JSX.Element | null>;
-export default _default;
