+++ generatedTypes/src/incubator/Dialog/HeaderContent.d.ts
